requirements.txt
python-telegram-bot==20.7
gspread==5.12.0
oauth2client==4.1.3
python-dotenv==1.0.0